downloadStream: video-payer-buttons.js -> download-stream.js
